pyproject.toml
[tool.poetry]
name = "puggle"
version = "0.1.5"
description = ""
authors = ["Michael Stewart <michael.stewart.webdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
colorama = "^0.4.6"
py2neo = "^2021.2.4"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]


[tool.poetry.group.test.dependencies]
pytest-cov = "^4.1.0"
coverage-badge = "^1.1.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
